+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This project follows [Semantic Versioning](https://semver.org/) and the [Keep a 
 - #51: ensured all boolean fields in AST nodes are initialised
 - #77: addressed no return type warnings in lexer and parser
 
-## [0.4.2-alpha] - 08/08/2025
+## [0.5.0-alpha] - 08/08/2025
 ### Added
 - #53: added language support for `==` and `!=` operators
 - #68: added first pass of quantum simulator